requirements.txt
discord.py
gspread
python-dotenv
